+++ client/rpc/src/test/kotlin/net/corda/client/rpc/RPCPerformanceTests.kt
@@ -1,9 +1,8 @@
 package net.corda.client.rpc
 
 import com.google.common.base.Stopwatch
-import net.corda.client.rpc.internal.RPCClientConfiguration
+import net.corda.client.rpc.internal.CordaRPCClientConfigurationImpl
 import net.corda.core.messaging.RPCOps
-import net.corda.core.utilities.days
 import net.corda.core.utilities.minutes
 import net.corda.core.utilities.seconds
 import net.corda.node.services.messaging.RPCServerConfiguration
@@ -43,7 +42,7 @@ class RPCPerformanceTests : AbstractRPCTest() {
     }
 
     private fun RPCDriverDSL.testProxy(
-            clientConfiguration: RPCClientConfiguration,
+            clientConfiguration: CordaRPCClientConfigurationImpl,
             serverConfiguration: RPCServerConfiguration
     ): TestProxy<TestOps> {
         return testProxy<TestOps>(
@@ -56,7 +55,7 @@ class RPCPerformanceTests : AbstractRPCTest() {
     private fun warmup() {
         rpcDriver {
             val proxy = testProxy(
-                    RPCClientConfiguration.default,
+                    CordaRPCClientConfigurationImpl.default,
                     RPCServerConfiguration.default
             )
             val executor = Executors.newFixedThreadPool(4)
@@ -86,8 +85,7 @@ class RPCPerformanceTests : AbstractRPCTest() {
         measure(inputOutputSizes, (1..5)) { inputOutputSize, _ ->
             rpcDriver {
                 val proxy = testProxy(
-                        RPCClientConfiguration.default.copy(
-                                cacheConcurrencyLevel = 16,
+                        CordaRPCClientConfigurationImpl.default.copy(
                                 observationExecutorPoolSize = 2
                         ),
                         RPCServerConfiguration.default.copy(
@@ -126,9 +124,8 @@ class RPCPerformanceTests : AbstractRPCTest() {
         rpcDriver {
             val metricRegistry = startReporter(shutdownManager)
             val proxy = testProxy(
-                    RPCClientConfiguration.default.copy(
-                            reapInterval = 1.seconds,
-                            cacheConcurrencyLevel = 16
+                    CordaRPCClientConfigurationImpl.default.copy(
+                            reapInterval = 1.seconds
                     ),
                     RPCServerConfiguration.default.copy(
                             rpcThreadPoolSize = 8
@@ -159,7 +156,7 @@ class RPCPerformanceTests : AbstractRPCTest() {
             // TODO this hangs with more parallelism
             rpcDriver {
                 val proxy = testProxy(
-                        RPCClientConfiguration.default,
+                        CordaRPCClientConfigurationImpl.default,
                         RPCServerConfiguration.default
                 )
                 val numberOfMessages = 1000