+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/config/ConfigUtilities.kt
@@ -9,6 +9,7 @@ import net.corda.core.internal.uncheckedCast
 import net.corda.core.utilities.NetworkHostAndPort
 import org.slf4j.LoggerFactory
 import java.lang.reflect.Field
+import java.lang.reflect.InvocationTargetException
 import java.lang.reflect.Modifier.isStatic
 import java.lang.reflect.ParameterizedType
 import java.net.Proxy
@@ -62,19 +63,20 @@ fun <T : Any> Config.parseAs(clazz: KClass<T>): T {
                 val path = defaultToOldPath(property)
                 getValueInternal<Any>(path, param.type)
             }
-    return constructor.callBy(args)
+    try {
+        return constructor.callBy(args)
+    } catch (e: InvocationTargetException) {
+        throw e.cause!!
+    }
 }
 
 class UnknownConfigurationKeysException private constructor(val unknownKeys: Set<String>) : IllegalArgumentException(message(unknownKeys)) {
-
     init {
         require(unknownKeys.isNotEmpty()) { "Absence of unknown keys should not raise UnknownConfigurationKeysException." }
     }
 
     companion object {
-
         fun of(offendingKeys: Set<String>): UnknownConfigurationKeysException = UnknownConfigurationKeysException(offendingKeys)
-
         private fun message(offendingKeys: Set<String>) = "Unknown configuration keys: ${offendingKeys.joinToString(", ", "[", "]")}."
     }
 }