+++ node/src/main/kotlin/net/corda/node/internal/AbstractNode.kt
@@ -86,6 +86,7 @@ import java.time.Duration
 import java.util.*
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ExecutorService
+import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit.SECONDS
 import kotlin.collections.set
 import kotlin.reflect.KClass
@@ -145,6 +146,8 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
     protected lateinit var networkMapUpdater: NetworkMapUpdater
     lateinit var securityManager: RPCSecurityManager
 
+    private val shutdownExecutor = Executors.newSingleThreadExecutor()
+
     /** Completes once the node has successfully registered with the network map service
      * or has loaded network map data from local database */
     val nodeReadyFuture: CordaFuture<Unit> get() = _nodeReadyFuture
@@ -159,7 +162,8 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
 
     /** The implementation of the [CordaRPCOps] interface used by this node. */
     open fun makeRPCOps(flowStarter: FlowStarter, database: CordaPersistence, smm: StateMachineManager): CordaRPCOps {
-        return SecureCordaRPCOps(services, smm, database, flowStarter)
+
+        return SecureCordaRPCOps(services, smm, database, flowStarter, { shutdownExecutor.submit { stop() } })
     }
 
     private fun initCertificate() {
@@ -302,7 +306,7 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
             if (configuration.rpcOptions.address == null) {
                 throw ConfigurationException("Cannot init CrashShell because node RPC address is not set (via 'rpcSettings' option).")
             }
-            InteractiveShell.startShell(configuration.toShellConfig())
+            InteractiveShell.startShell(configuration.toShellConfig(), cordappLoader.appClassLoader)
         }
     }
 
@@ -673,7 +677,7 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
         val notaryKey = myNotaryIdentity?.owningKey ?: throw IllegalArgumentException("No notary identity initialized when creating a notary service")
         return notaryConfig.run {
             if (raft != null) {
-                val uniquenessProvider = RaftUniquenessProvider(configuration, database, services.monitoringService.metrics, raft)
+                val uniquenessProvider = RaftUniquenessProvider(configuration, database, services.clock, services.monitoringService.metrics, raft)
                 (if (validating) ::RaftValidatingNotaryService else ::RaftNonValidatingNotaryService)(services, notaryKey, uniquenessProvider)
             } else if (bftSMaRt != null) {
                 if (validating) throw IllegalArgumentException("Validating BFTSMaRt notary not supported")
@@ -712,6 +716,7 @@ abstract class AbstractNode(val configuration: NodeConfiguration,
             toRun()
         }
         runOnStop.clear()
+        shutdownExecutor.shutdown()
         _started = null
     }