+++ node/src/main/kotlin/net/corda/node/services/transactions/ValidatingNotaryFlow.kt
@@ -5,12 +5,11 @@ import net.corda.core.contracts.TimeWindow
 import net.corda.core.contracts.TransactionVerificationException
 import net.corda.core.flows.*
 import net.corda.core.internal.ResolveTransactionsFlow
-import net.corda.core.internal.validateRequest
+import net.corda.core.internal.validateRequestSignature
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionWithSignatures
 import net.corda.core.transactions.WireTransaction
-import net.corda.core.utilities.unwrap
 import java.security.SignatureException
 
 /**
@@ -25,9 +24,10 @@ class ValidatingNotaryFlow(otherSideSession: FlowSession, service: TrustedAuthor
      * the transaction in question has all required signatures apart from the notary's.
      */
     @Suspendable
-    override fun receiveAndVerifyTx(): TransactionParts {
+    override fun validateRequest(requestPayload: NotarisationPayload): TransactionParts {
         try {
-            val stx = receiveTransaction()
+            val stx = requestPayload.signedTransaction
+            validateRequestSignature(NotarisationRequest(stx.inputs, stx.id), requestPayload.requestSignature)
             val notary = stx.notary
             checkNotary(notary)
             resolveAndContractVerify(stx)
@@ -40,15 +40,6 @@ class ValidatingNotaryFlow(otherSideSession: FlowSession, service: TrustedAuthor
                 is SignatureException -> NotaryInternalException(NotaryError.TransactionInvalid(e))
                 else -> e
             }
-        }
-    }
-
-    @Suspendable
-    private fun receiveTransaction(): SignedTransaction {
-        return otherSideSession.receive<NotarisationPayload>().unwrap {
-            val stx = it.signedTransaction
-            validateRequest(NotarisationRequest(stx.inputs, stx.id), it.requestSignature)
-            stx
         }
     }